service: Move suite definitions to id_test.go
We no longer have multiple implementations of the suite, so
it's fine to move to the same module as the tests are located in.
