--- pkg/service/service_test.go
@@ -1,44 +1,0 @@
-// Copyright 2018 Authors of Cilium
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-// +build !privileged_tests
-
-package service
-
-import (
-	"testing"
-
-	. "gopkg.in/check.v1"
-)
-
-// Hook up gocheck into the "go test" runner.
-func Test(t *testing.T) {
-	TestingT(t)
-}
-
-type ServiceTestSuite struct{}
-
-type ServiceLocalSuite struct {
-	ServiceTestSuite
-}
-
-var _ = Suite(&ServiceLocalSuite{})
-
-func (e *ServiceLocalSuite) SetUpTest(c *C) {
-	serviceIDAlloc.resetLocalID()
-}
-
-func (e *ServiceLocalSuite) TearDownTest(c *C) {
-	serviceIDAlloc.resetLocalID()
-}
